move announcement code to hostannounce.go
